% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chulldata.R
\docType{data}
\name{chulldata}
\alias{chulldata}
\title{Reduced k-means on iris data}
\format{
An object of class \code{data.frame} with 14 rows and 2 columns.
}
\usage{
data(chulldata)
}
\description{
Reduced k-means on iris data
}
\examples{
data(chulldata)
}
\keyword{datasets}
