% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chulldataboot.R
\docType{data}
\name{chulldataboot}
\alias{chulldataboot}
\title{Reduced k-means on bootstrapped iris data}
\format{
An object of class \code{data.frame} with 14 rows and 101 columns.
}
\usage{
data(chulldataboot)
}
\description{
Reduced k-means on bootstrapped iris data
}
\examples{
data(chulldataboot)
}
\keyword{datasets}
